+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 * Ensure emails have name, destination and subject
 * Fix bug in TP account deletion
 * UI improvements on projects page
+* Fix database creation validation confirmation email
+* Fix database creation request email title
+* Remove DB 'size' attribute
 
 ## 1.4.32 (2025-04-22)